MOST-up-to-date-ai/requirements.txt
# Core Dependencies
opencv-python>=4.8.0.74
numpy>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
google-api-python-client>=2.90.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Required for testing
pytest-mock>=3.11.10
